pyproject.toml
[project]
name = "echo-line-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "line-bot-sdk>=3.17.1",
    "openai>=1.0.0",
    "boto3>=1.28.0",
]
